+++ Verification/Consumer/Package.swift
@@ -15,7 +15,7 @@ let package = Package(
     dependencies: [
         .package(
             url: "https://github.com/SunsetWan/DeckStringDecoderKMPPackage.git",
-            exact: "0.1.0-kmp.5"
+            exact: "0.1.0-kmp.6"
         ),
     ],
     targets: [
@@ -29,7 +29,6 @@ let package = Package(
             name: "DeckStringDecoderPublicConsumerTests",
             dependencies: [
                 "DeckStringDecoderPublicConsumer",
-                .product(name: "DeckStringDecoder", package: "DeckStringDecoderKMPPackage"),
             ]
         ),
     ]